app/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
opencv-python
python-multipart